Don't scale boundary lines with zoom
This means that when as we zoom in the boundary lines will remain at a constant thickness to the viewer, rather than scaling up.  The advantage is that we can zoom in on smaller constituencies (like those in cities) from the large maps and they don't disappear in a blob of boundary lines.

See: http://www.w3.org/TR/SVGTiny12/painting.html#NonScalingStroke