Move #constituency_info management to form.js
Now map.js is responsible for drawing the constituency on the map and form.js is responsible for drawing the constituency info on the UI.  We use events to trigger the changes and so map.js doesn't need to know about form.js and vice-versa.

This means we can reduce the API exposed on PetitionMap by removing selectConstituency and moving numberWithCommas into form.js.