Add changes since May 2015 General Election
 * Glasgow East and Edinburgh West candidates now "Independent", not "SNP"
 * Oldham West and Royton by-election result

This file was generated by extracting the data from http://data.parliament.uk/membersdataplatform/services/mnis/members/query/House=Commons%7CIsEligible=true/ and condensing it down to just mp, party, and constituency names then matching the names in uk/topo_wpc.json to get the ons_code.

It's not clear how we generated the original version of this file, so we'll want to come up with a better way to regenerate it to keep up with future changes.