Prefer fixture data for filling in constituency panel
When filling in the constituency info panel we were getting the names of the MP and constituency from the petition API if there were signatures in that constituency, but from our fixture data if there weren't.  It's better to use a single data source if we can because then we don't run the risk of discrepancies in the data.  Ideally we'd use the petition API for everything as the data there comes direct from the parliament API and will always be uptodate.  Unfortunately that would require a couple of changes; include all constituencies even if there were no votes, include the party of the MP.  The first would be easy, but the second would require large changes to how the petition app fetches constituency/mp info because the current parliament API method used doesn't include the part information.

It's unlikely that the tail of the map would wag the dog of the petition site so we'll just always use our fixture data and pay attention to keeping it updated regularly.  Happily, this simplifies the code a little anyway, so that's nice.