Define CSS styles for areas in the 0th bucket
When no-one in a constituency signs a petition usually the constituency does
not appear in the JSON signatures_by_constituency payload so we did not need
to have a CSS selector for entries in the 0th bucket because we wouldn't assign
anything to it.  However, in rare cases (fraud removal usually) the constituency
can appear in the JSON and without a CSS selector they get a full black fill.

We also remove the t1-t8 classes used for the key, as we can re-use the c0-c8
classes here and not run the risk of the colouring for the key and
constituencies getting out of sync.