Add HexJSON representation of constituencies
This file represents every UK parliamentary constituency as a hexagon
arranged on a grid that maintains a semblance of geography.  This will let
us show a different view of the UK that won't be overwhelmed by all the
sparsely populated rural constituencies and hide all the densely populated
urban constituencies.  The data was created by the [ODI Leeds][odi-leeds]
as part of their [Hexmaps project][odi-hexmaps], thanks to them for making
the data open; it makes things better!

[odi-leeds]: https://odileeds.org
[odi-leeds-hexmaps]: https://odileeds.org/projects/hexmaps/constituencies/