Pass constituency id, not constituency object in event
The object we send to the events is the constituency representation from
the topo json file.  The structure of this object will be different when
we introduce the hex json version.  As it happens we only use the id of
the constituency in any of our handlers, and this will be present on the
hex json objects, so we can just pass that instead.