Introduce `PetitionMap.mapMode` and draw some hexagons
If the mapMode is set to `topo` (the default) we draw the geographic maps
we always used to.  If we set it to `hex` though, we'll draw the hexagon
map based on the hexJSON file from the ODI.  We don't need the projection
and path stuff from the topo json maps because those are to do with
translating spherical maps onto a flat plane.  Other than that though, it
is pretty straightforward.  Our hex rendering is based pretty strongly on
the examples from the [d3-hexjson project][d3-hexjson].

[d3-hexjson]: https://github.com/olihawkins/d3-hexjson/