Add hexjson files for map subdivisions
Originally I hoped to just filter the uk map to only include the relevant
constituencies, but for some reason the d3-hexjson library would render
some of the files incorrectly.  The problem seems to be that the reduced
number of hexagons in each file means we have to translate the `q` and `r`
attributes of each hexagon into an absolute grid reference, and sometimes
this means we move a hexagon from an odd or even `q` or `r` onto an even
or odd one.  As [rendering hexagons by grid reference][hex-coords] is
dependent on the odd and even state of the `q` and `r` attributes, this
meant that for some of the smaller maps some rows were drawn incorrectly,
making the whole map wrong.

I tried detecting when the absolute `q` and `r` had a different odd or
even state compared to the original, but this meant that the rendering of
the map was offset by a row or column, and so the arrangement was good,
but it wasn't centered on the screen.

The solution I ended up with was to manually translate the `q` and `r`
attributes of each map to an absolute space that preserved the correct
odd and even states and didn't cause a problem.

This was painful, but perhaps could be automated later if we need to
regenerate the subdivisions for any future constituency changes.

[hex-coords]: https://www.redblobgames.com/grids/hexagons/#coordinates